refactor(kimi): centralize delegated agent activity and task ownership
